Added grey transparent image and began editiing headshot title
I added the image that adds the grey transparent filter on the buttons for my portfolio when because I didn't add it initially to the repo. I began editing headshot title so that my headshot is bigger, my name is next to my headshot, and I plan on having my titles below my name and headshot or they'll go where they were before.